Update RandomizedQueue.java
random iterator works. I think I can improve the speed of the dequeue method by implememnting a private swap method that interchanges the values at two different indices in the array. So I can randomly select an index, and then swap that index with the terminal entry in the array, and then remove the terminal entry in the array.