Update BruteCollinearPoints.java
changed return type of fourTuples to an ArrayList with nested Arrays (instead of ArrayList with nested ArrayLists). I did this because the function explicitly returns only arrays of size 4. And because of this, we can set the size of each inner array to size four. This saves us from having to convernt ArrayLists into arrays further down the road.